Reorganize UI widgets into dedicated feature folders
- Move chat components to ui/chat/: chat_widget, chat_ai_item, chat_human_item, chat_item_mixin, chat_tool_item
- Move input widget to ui/input/: input_widget
- Move status widget to ui/status/: status_widget
- Update all import references in main_window.py to reflect new structure
- Remove unused ToolMessage import to maintain code quality
- Improve code organization and maintainability through better separation of concerns
