+++ speedoflight/ui/chat/chat_ai_item.py
@@ -4,7 +4,7 @@ from gi.repository import Adw, Gtk, Pango  # type: ignore
 
 from speedoflight.constants import DEFAULT_MARGIN
 from speedoflight.models import GBaseMessage
-from speedoflight.ui.main.chat_item_mixin import ChatItemMixin
+from speedoflight.ui.chat.chat_item_mixin import ChatItemMixin
 
 
 class AIMessageWidget(Gtk.Box, ChatItemMixin):
