+++ speedoflight/ui/status/status_widget.py
